# backend/app/services/chatgpt_plugin/openapi.yaml
openapi: 3.0.1
info:
  title: Yasmin AI Builder API
  description: Build apps, websites, games, and systems with parallel AI agents
  version: 'v1'
  contact:
    name: Yasmin Support
    email: support@yasmin.ai
servers:
  - url: https://api.yasmin.ai/v1
paths:
  /projects:
    get:
      operationId: listProjects
      summary: List all projects
      description: Get a list of all projects in your Yasmin account
      parameters:
        - name: status
          in: query
          description: Filter by project status
          schema:
            type: string
            enum: [draft, active, archived]
        - name: type
          in: query
          description: Filter by project type
          schema:
            type: string
            enum: [web, mobile, game, api, system]
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
    post:
      operationId: createProject
      summary: Create a new project
      description: Create a new project with AI agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get project details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      operationId: deleteProject
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted

  /projects/{projectId}/agents:
    get:
      operationId: listAgents
      summary: List agents in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
    post:
      operationId: createAgent
      summary: Create a new agent
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        "201":
          description: Agent created

  /projects/{projectId}/deploy:
    post:
      operationId: deployProject
      summary: Deploy a project
      description: Deploy the project to production or staging
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
      responses:
        "202":
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'

  /agents/{agentId}/run:
    post:
      operationId: runAgent
      summary: Execute an agent task
      description: Run a specific agent with input parameters
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunAgentRequest'
      responses:
        "200":
          description: Agent execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResult'

  /code/generate:
    post:
      operationId: generateCode
      summary: Generate code using AI
      description: Generate code for a specific framework or language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCodeRequest'
      responses:
        "200":
          description: Generated code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedCode'

  /deployments:
    get:
      operationId: listDeployments
      summary: List deployments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed]
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentList'

components:
  schemas:
    Project:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        type: {type: string, enum: [web, mobile, game, api, system]}
        status: {type: string, enum: [draft, active, archived]}
        owner_id: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        agents_count: {type: integer}
        deployments_count: {type: integer}

    ProjectList:
      type: object
      properties:
        projects: {type: array, items: {$ref: '#/components/schemas/Project'}}
        total: {type: integer}
        page: {type: integer}

    CreateProjectRequest:
      type: object
      required: [name, type]
      properties:
        name: {type: string, description: "Project name"}
        description: {type: string}
        type: {type: string, enum: [web, mobile, game, api, system]}
        template: {type: string, description: "Optional project template"}
        config: {type: object}

    Agent:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        role: {type: string}
        status: {type: string, enum: [idle, running, error, paused]}
        project_id: {type: string}
        capabilities: {type: array, items: {type: string}}
        last_active: {type: string, format: date-time}

    AgentList:
      type: object
      properties:
        agents: {type: array, items: {$ref: '#/components/schemas/Agent'}}

    CreateAgentRequest:
      type: object
      required: [name, role]
      properties:
        name: {type: string}
        role: {type: string, description: "Agent role: frontend, backend, devops, designer, tester, etc."}
        capabilities: {type: array, items: {type: string}}
        config: {type: object}

    DeployRequest:
      type: object
      required: [environment]
      properties:
        environment: {type: string, enum: [staging, production]}
        version: {type: string}
        config: {type: object}

    Deployment:
      type: object
      properties:
        id: {type: string}
        project_id: {type: string}
        status: {type: string}
        environment: {type: string}
        url: {type: string}
        started_at: {type: string}
        completed_at: {type: string}

    DeploymentList:
      type: object
      properties:
        deployments: {type: array, items: {$ref: '#/components/schemas/Deployment'}}

    RunAgentRequest:
      type: object
      required: [task]
      properties:
        task: {type: string, description: "Description of the task for the agent"}
        parameters: {type: object}
        timeout: {type: integer, default: 300}

    AgentResult:
      type: object
      properties:
        success: {type: boolean}
        output: {type: string}
        artifacts: {type: array, items: {type: object}}
        execution_time: {type: number}

    GenerateCodeRequest:
      type: object
      required: [description, framework]
      properties:
        description: {type: string, description: "What code to generate"}
        framework: {type: string, enum: [react, vue, angular, flutter, node, python, unity, unreal]}
        language: {type: string}
        context: {type: string}

    GeneratedCode:
      type: object
      properties:
        files: {type: array, items: {type: object}}
        explanation: {type: string}
        dependencies: {type: array, items: {type: string}}